separate pojo and service
